requirements.txt
streamlit==1.29.0
fpdf
PyPDF2
Pillow
pytesseract
newspaper3k
pandas
